+++ examples/vue-apollo/apollo/network-interfaces/default.js
@@ -1,5 +1,0 @@
-import { createNetworkInterface } from 'apollo-client'
-
-export default (ctx) => {
-  return createNetworkInterface({ uri: 'https://api.graph.cool/simple/v1/cj1dqiyvqqnmj0113yuqamkuu' })
-}