+++ examples/with-buefy/nuxt.config.js
@@ -3,8 +3,9 @@ export default {
   ** Customize the progress bar color
   */
   loading: { color: '#3B8070' },
-  mode: 'spa',
-  plugins: ['~plugins/buefy'],
+  modules: [
+    'nuxt-buefy'
+  ],
   router: {
     linkActiveClass: 'is-active'
   }